Rename auth API filenames
