+++ generators/app/templates/api/auth/controller.js
