+++ generators/app/templates/api/auth/index.test.js
