--- generators/app/templates/api/auth/package.json
@@ -1,6 +1,0 @@
-{
-  "name": "auth",
-  "version": "0.0.0",
-  "private": true,
-  "main": "./auth.router.js"
-}
