+++ generators/app/templates/api/password-reset/controller.js
