+++ generators/app/templates/api/password-reset/index.test.js
