+++ generators/app/templates/api/password-reset/model.js
