+++ generators/app/templates/api/password-reset/model.test.js
