+++ generators/app/templates/api/user/controller.js
