+++ generators/app/templates/api/user/model.js
