+++ generators/app/templates/api/user/model.test.js
