--- generators/app/templates/api/user/package.json
@@ -1,6 +1,0 @@
-{
-  "name": "user",
-  "version": "0.0.0",
-  "private": true,
-  "main": "./user.router.js"
-}
